+++ src/examples/ec_gravity_compensation/src/ec_gravity_compensation.cpp
@@ -95,7 +95,7 @@ int main(int argc, char * const argv[])
         uint64_t start_time_ns = iit::ecat::get_time_ns();
         uint64_t time_ns=start_time_ns;
         
-        double time_elapsed_ms;
+        float time_elapsed_ms;
         
         bool first_Rx=false;
 
@@ -116,7 +116,7 @@ int main(int argc, char * const argv[])
             imp_zero_gains_map[id][2]=0.0;
         }
 
-        double set_gain_time_ms=3000; // Default 3s 
+        float set_gain_time_ms=3000; // Default 3s 
         std::map<int,std::vector<float>> gain_start_map=imp_gains_map;
         std::map<int,std::vector<float>> gain_trj_map=imp_zero_gains_map;
         std::map<int,std::vector<float>> gain_ref_map=gain_start_map;
@@ -133,7 +133,7 @@ int main(int argc, char * const argv[])
         uint32_t cmd_aux_sts,brake_sts,led_sts;
         MotorStatusMap motors_status_map;
         
-        double tau=0,alpha=0;
+        float tau=0,alpha=0;
         MotorReferenceMap motors_ref;
         
         // memory allocation
@@ -204,7 +204,7 @@ int main(int argc, char * const argv[])
         
         while (run_loop && client->is_client_alive())
         {
-            time_elapsed_ms= (time_ns-start_time_ns)/1000000;
+            time_elapsed_ms= (static_cast<float>((time_ns-start_time_ns))/1000000);
             //DPRINTF("Time [%f]\n",time_elapsed_ms);
             
             // Rx "SENSE"
